backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-api-python-client==2.97.0
google-cloud-aiplatform==1.38.1
python-dotenv==1.1.0
requests==2.31.0
youtube-transcript-api==0.6.1
firebase-admin==6.2.0
Flask-SQLAlchemy==3.1.1
